+++ R/IMAGE_PWMs.R
@@ -1,3 +1,0 @@
-IMAGE_PWMs <- function(){
-    return(IMAGE_PWMlist)
-}