+++ R/reexports.R
@@ -1,0 +1,32 @@
+#' Reexports
+#' 
+#' reexport of functions
+#' 
+#' @name Reexports
+#' 
+#' @importFrom methylumi betas methylated unmethylated pval.detect<-
+#' @importFrom minfi getAnnotationObject
+#' @importMethodsFrom Biostrings reverseComplement width matchPWM start
+#' 
+#' @usage NA
+#' 
+#' @export
+minfi::getAnnotationObject
+#' @export
+methylumi::betas
+#' @export
+methylumi::methylated
+#' @export
+methylumi::unmethylated
+#' @export
+methylumi::"pval.detect<-"
+#' @export
+Biostrings::reverseComplement
+#' @export
+Biostrings::width
+#' @export
+Biostrings::matchPWM
+#' @export
+Biostrings::start
+
+